Additional xumm integration, to supply push notifications.  To use xumm, a redis server must be available right now as it stores the issued_user_token there.
